MicroservicioReportes/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
python-dotenv
pandas
openpyxl
reportlab
pymysql
